refactor: Separate SearchDbHelper
- Make SearchRepository
